update api url and keys in func updateOverAll()
